Restructure matrix files
