+++ lib/honey_format/matrix/row.rb
