+++ lib/honey_format/matrix/row_builder.rb
