+++ lib/honey_format/matrix/rows.rb
